+++ src/SoftTissue.Core/Operations/QuasiLinearViscoelasticity/CalculateStressSensitivityAnalysis/FungModel/ICalculateFungModelStressSentivityAnalysis.cs
@@ -1,6 +1,7 @@
 ﻿using SoftTissue.Core.Models.Viscoelasticity.QuasiLinear;
+using SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStressSensitivityAnalysis;
 
-namespace SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStress.FungModel
+namespace SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStressSensitivityAnalysis.FungModel
 {
     /// <summary>
     /// It is responsible to do a semsitivity analysis while calculating the stress to Fung model.